ai-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
anthropic==0.39.0
pydantic==2.9.0
python-dotenv==1.0.1
pytest==8.3.0
httpx==0.27.0
